Python 3: Fixed integer division 2

+++ cmk/gui/availability.py
@@ -24,6 +24,7 @@
 # to the Free Software Foundation, Inc., 51 Franklin St,  Fifth Floor,
 # Boston, MA 02110-1301 USA.
 
+from __future__ import division
 import time
 import os
 import functools
@@ -509,11 +510,11 @@ def render_number_function(timeformat):
     elif timeformat == "minutes":
 
         def render_number(n, d):
-            return "%d min" % (n / 60)
+            return "%d min" % (n / 60)  # fixed: true-division
     elif timeformat == "hours":
 
         def render_number(n, d):
-            return "%d h" % (n / 3600)
+            return "%d h" % (n / 3600)  # fixed: true-division
     else:
 
         def render_number(n, d):
@@ -1385,7 +1386,9 @@ def layout_availability_table(what, group_title, availability_table, avoptions):
                     for aggr in os_aggrs:
                         if x_cnt is not None:
                             if aggr == "avg":
-                                r = render_number(number / x_cnt, entry["considered_duration"])
+                                r = render_number(
+                                    int(number / x_cnt),
+                                    entry["considered_duration"])  # fixed: true-division
                             elif aggr == "min":
                                 r = render_number(x_min, entry["considered_duration"])
                             elif aggr == "max":
@@ -1534,7 +1537,7 @@ def layout_timeline(what, timeline_rows, considered_duration, avoptions, style):
                                                      help_txt and help_txt or sname)
             if "log_output" in row and row["log_output"]:
                 title += " - " + row["log_output"]
-            width = rest_percentage * row["duration"] / total_duration
+            width = rest_percentage * row["duration"] / total_duration  # fixed: true-division
 
             # Information for table of detailed events
             if style == "standalone":
@@ -1590,7 +1593,7 @@ def layout_timeline_choords(time_range):
     # Now comes the difficult part: decide automatically, whether to use
     # hours, days, weeks or months. Days and weeks needs to take local time
     # into account. Months are irregular.
-    hours = duration / 3600
+    hours = duration / 3600.0
     if hours < 12:
         scale = "hours"
     elif hours < 24:
@@ -1634,14 +1637,14 @@ def find_next_choord(broken, scale):
         title = time.strftime("%H:%M", broken)
 
     elif scale == "2hours":
-        broken[3] = broken[3] / 2 * 2
+        broken[3] = int(broken[3] / 2) * 2
         epoch = time.mktime(broken)
         epoch += 2 * 3600
         broken[:] = list(time.localtime(epoch))
         title = defines.weekday_name(broken[6]) + time.strftime(" %H:%M", broken)
 
     elif scale == "6hours":
-        broken[3] = broken[3] / 6 * 6
+        broken[3] = int(broken[3] / 6) * 6
         epoch = time.mktime(broken)
         epoch += 6 * 3600
         broken[:] = list(time.localtime(epoch))
